Add options to toggle off components
With the addition of the `primary-component-toggled-off` and
`secondary-component-toggled-off` properties, it's not possible to
toggle off the left or right (or top or bottom) components.

Note that attempting to toggle both of them off results in an exception
being thrown, so the user of the library should take care that it
doesn't happen.

Closes: wannabegeek/ng2-split-pane#10